Generating build files with cmake (#21)
* ignoring cmake related build files

* Adding LLVM dependency as external dependecy.

Using git submodules.

* Initial workspace setup with CMake

* Including tests into the CMake build.

Adding the LLVM module so that all `lit` related
functions are available.

A new target is made available for running tests: `check-mlir-tutorial`.

* Including tools directory into the CMake build

So that `tutorial-opt` can be built. This introduces as well
a new target `tutorial-opt`.

* Adding affine full unroll related passes to cmake build

This introduces a new mlir lib: `AffineFullUnroll` and a new
build target to build it: `MLIRAffineFullUnrollPasses`.

* Adding tablgen cmake module

to have access to all tablgen related cmake statements.

* add mlir include dirs

* Adding mul to add related passes to cmake build

This introduces a new mlir lib: `MulToAdd` and a new
build target to build it: `MLIRMulToAddPasses`.

* Renaming `PolyOps.td` to `Poly.td`

So that on the cmake side of things we can use the `add_mlir_dialect`
macro without too many changes to the include directives.

Ref: https://github.com/llvm/llvm-project/blob/main/mlir/cmake/modules/AddMLIR.cmake#L166-L177

* ignoring vs code metafiles plus compilation database files.

* Adding poly dialect lib to cmake build

* Adding CI for cmake build.

* Adding canonicalizers to cmake build

* Inlining cmake commands for Poly dialect generation.

Instead of using `add_mlir_dialect` helper function for dialect generation,
we inline it here so that we can have control over the naming of
generated `*.inc` files.

This avoids renaming of `PolyOps.td` so that cmake builds work
well on the existing source code.